+++ @navikt/internal/react/src/dropdown-menu/index.ts
@@ -1,4 +1,2 @@
 export { default as DropdownMenu } from "./DropdownMenu";
 export * from "./DropdownMenu";
-export { default as DropdownMenuItem } from "./Item";
-export * from "./Item";
